% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brdwgd_weather.R
\name{get_brdwgd_weather}
\alias{get_brdwgd_weather}
\title{Extract Point Data from BR-DWGD}
\usage{
get_brdwgd_weather(
  target_longitude,
  target_latitude,
  weather_variables = c("pr", "Tmin", "Tmax", "Rs", "RH", "u2", "ETo"),
  date_range = NULL,
  nc_files_directory = "."
)
}
\arguments{
\item{target_longitude}{Numeric. Longitude of the target location in decimal degrees.}

\item{target_latitude}{Numeric. Latitude of the target location in decimal degrees.}

\item{weather_variables}{Character vector. Variables to extract. Allowed values are: \code{"pr"} (precipitation), \code{"Tmin"}, \code{"Tmax"}, \code{"Rs"} (solar radiation), \code{"RH"} (relative humidity), \code{"u2"} (wind speed), and \code{"ETo"}. Defaults to all.}

\item{date_range}{Character vector of length 2 (\code{"YYYY-MM-DD"}). The start and end dates. If \code{NULL} (default), extracts the entire available period.}

\item{nc_files_directory}{Character. Path to the local directory containing the BR-DWGD NetCDF files. Defaults to current working directory \code{"."}.}
}
\value{
A \code{tibble} containing the \code{date}, the actual \code{lat} and \code{lon} of the closest grid cell, and columns for each requested weather variable.
}
\description{
Extracts a daily time series of specified weather variables for a given longitude and latitude
from the Brazilian Daily Weather Gridded Data (BR-DWGD) NetCDF files. The function identifies
the closest grid cell to the target coordinates.
}
\examples{
target_longitude = -50.5995
target_latitude = -27.2863
weather_variables = c("pr", "ETo", "Tmax")
date_range = c("1983-01-01", "1983-03-31")
nc_files_directory = "D:/clima_Xavier"


\dontrun{
# Ensure the path to your NetCDF files is correct
weather_data_point <- get_brdwgd_weather(
  target_longitude = -50.5995,
  target_latitude = -27.2863,
  weather_variables = c("pr", "ETo", "Tmax"),
  date_range = c("1983-01-01", "1983-03-31"),
  nc_files_directory = "D:/clima_Xavier"
)
print(head(weather_data_point))
}
}
\references{
Xavier, A. C., King, C. W., & Scanlon, B. R. (2016). Daily gridded meteorological variables in Brazil (1980–2013). International Journal of Climatology. \url{https://sites.google.com/site/alexandrecandidoxavierufes/brazilian-daily-weather-gridded-data}
}
