% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_unit.R
\name{flow_unit}
\alias{flow_unit}
\title{Flow unit conversion}
\usage{
flow_unit(
  flow_rate,
  flow_unit = c("m3/s", "l/h", "m3/h", "l/s"),
  operator = c("div", "mult")
)
}
\arguments{
\item{flow_rate}{Flow rate value.}

\item{flow_unit}{Flow measurement unit. Options are "m3/s", "l/h", "m3/h", or "l/s". Default is "m3/s".}

\item{operator}{Character indicating the operation: "div" to convert to m3/s, "mult" to convert from m3/s.}
}
\value{
Converted flow rate.
}
\description{
Helper function to convert flow rates between different units.
}
\keyword{internal}
